remove window buttons
