+++ src-tauri/Cargo.toml
@@ -16,7 +16,7 @@ tauri-build = { version = "1.5.5", features = [] }
 [dependencies]
 serde_json = "1.0"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "1.8.1", features = ["api-all"] }
+tauri = { version = "1.8.1", features = [ "macos-private-api", "api-all"] }
 vosk = "0.2.0"
 # cpal 0.15.3 has a bug that causes the app to crash on macOS with coreaudio-rs
 cpal = "0.14.1"
